changed program name to SuperQuickFood.java
